Armor
It is still early in development, but none-the-less, it is implemented. You still get the noise, and whatnot when you get hurt, while wearing the armor, but that should be fixe in the next commit. You also get flight from the pants... Should also be fixed in the next commit.